requirements.txt
numpy
pytorch
scipy
tqdm
wandb
scikit-learn
pandas
anndata
scanpy
gseapy